Organize foldre structure
