+++ src/components/shared/containers/DetailContent.jsx
