+++ src/pages/HomePage.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import logo from './assets/img/logo.svg';
+import logo from '../assets/img/logo.svg';
 
 const HomePage = () => (
   <div className="App">
