+++ src/pages/NormalPage.jsx
@@ -2,14 +2,11 @@
 import React from 'react';
 import { Switch, Route } from 'react-router-dom';
 // import PropTypes from 'prop-types';
-import DetailContent from '../components/shared/DetailContent';
-import ListContent from '../components/shared/ListContent';
-import MainContent from '../components/shared/MainContent';
-import Sidebar from '../components/shared/Sidebar';
-import Nav from '../components/shared/Nav';
-import Navbar from '../components/shared/Navbar';
-import logo from './assets/img/logo.png';
-import './assets/css/NormalPage.css';
+import DetailContent from '../components/shared/containers/DetailContent';
+import ListContent from '../components/shared/containers/ListContent';
+import MainContent from '../components/shared/containers/MainContent';
+import Sidebar from '../components/shared/children/Sidebar';
+import Nav from '../components/shared/children/Nav';
 
 const shoppingLists = [
   {
@@ -49,38 +46,35 @@ const linkArray = [
 
 /* eslint-disable react/prop-types */
 const NormalPage = () => (
-  <div className="lay_fluid-container">
-    <Navbar Navs={Nav} links={linkArray} siteTitle="ShoppingList" logo={logo} onlyLogo={false} />
-    <div className="lay_row lay_strict">
-      <Sidebar>
-        <Nav links={linkArray} extraClass="nav-sidebar" />
-      </Sidebar>
-      <MainContent>
-        <Switch>
-          <Route
-            exact
-            path="/normal"
-            render={props => (
-              <div className="table-responsive">
-                <h2 className="sub-header">Shopping Lists</h2>
-                <ListContent headerClass="page-header" objects={shoppingLists} {...props} />
-              </div>
-            )}
-          />
-          <Route
-            exact
-            path="/normal/:id"
-            render={props => (
-              <DetailContent
-                headerClass="page-header"
-                object={shoppingLists.find(list => list.id.toString() === props.match.params.id)}
-                {...props}
-              />
-            )}
-          />
-        </Switch>
-      </MainContent>
-    </div>
+  <div className="lay_row lay_strict">
+    <Sidebar>
+      <Nav links={linkArray} extraClass="nav-sidebar" />
+    </Sidebar>
+    <MainContent>
+      <Switch>
+        <Route
+          exact
+          path="/normal"
+          render={props => (
+            <div className="table-responsive">
+              <h2 className="sub-header">Shopping Lists</h2>
+              <ListContent headerClass="page-header" objects={shoppingLists} {...props} />
+            </div>
+          )}
+        />
+        <Route
+          exact
+          path="/normal/:id"
+          render={props => (
+            <DetailContent
+              headerClass="page-header"
+              object={shoppingLists.find(list => list.id.toString() === props.match.params.id)}
+              {...props}
+            />
+          )}
+        />
+      </Switch>
+    </MainContent>
   </div>
 );
 /* eslint-enable react/prop-types */
