--- src/pages/assets/css/styles.css
@@ -1,76 +1,0 @@
-.home-page {
-  /* background: url("../assets/img/shopping_carts.jpg") center center no-repeat;*/
-  min-height: 100%;
-  background-size: cover;
-}
-
-.navbar {
-  background-color: #266850;
-  color: #fff;
-}
-.navbar a {
-  color: #fff;
-}
-
-#no-background-image {
-  background: #fff;
-  min-height: 100%;
-}
-
-.front-page {
-  background: transparent;
-  color: #fff;
-}
-
-#front-page-dark {
-  background-color: #000;
-  background: transparent;
-  background-color: rgba(0, 0, 0, 0.6);
-  min-height: 92vh;
-}
-
-.front-page-content {
-  padding: 30vh 2vw 30vh 2vw;
-}
-
-#home-page-form {
-  background: #266850;
-  padding: 1px 0;
-}
-
-#front-form-tab-content {
-  padding: 5% 0;
-}
-
-.front-submit {
-  background-color: #266850;
-  color: #fff;
-  border: 1px solid #fff;
-}
-
-#home-page-form ul.nav-tabs li.nav-item a {
-  background-color: #fff;
-  color: #000;
-}
-
-#home-page-form ul.nav-tabs li.nav-item .active {
-  background-color: #266850;
-  color: #fff;
-  border-color: #266850;
-}
-
-.home-intro {
-  font-size: 1.3em;
-}
-
-#left-sidebar {
-  border-right-color: #266850;
-  border-right-width: 2px;
-  border-right-style: solid;
-}
-
-.sidebar_card {
-  border-bottom-color: #266850;
-  border-bottom-width: 2px;
-  border-bottom-style: solid;
-}
